net: phy: at803x: use phydev phy-mode in aneg_done
at803x_aneg_done currently reads the chip configuration register to
determine the operation mode of the AR8031/AR8033 PHY.

This read has to be performed every time the link-state changes. It's
also inconsistent, as some places already use the phy-mode defined for
the phydev.

Use the phy-mode of the phydev to remove this read and be more
consistent throughout the driver.

Signed-off-by: David Bauer <mail@david-bauer.net>