ath11k: add WoW quirk for ThinkPad X13 G4 (AMD)
The ThinkPad X13 Gen 4 (AMD) is sold with two model numbers, 21J3 and
21J4.

While the second was already in the list of devices needing the WoW
quirk to ensure the device does not wake from standby on AC state
change, the first model number was missing.

Add the 21J3 model to the list, fixing unintended standby wakes.

Fixes: ce8669a27016 ("wifi: ath11k: determine PM policy based on machine model")

Signed-off-by: David Bauer <mail@david-bauer.net>